🚚 (server): Move /pagination/userbookmarks folder
