+++ apps/server/src/pagination/userBookmarks/domain/errors/paginationUserBookmarks.errors.ts
